+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # Release Notes for Blitz
 
+## 4.10.1 - Unreleased
+
+### Fixed
+
+- Fixed a bug in which SQL queries were executed in the Blitz Diagnostics utility that required minimum database versions above those required by Craft 4 ([#602](https://github.com/putyourlightson/craft-blitz/issues/602)).
+
 ## 4.10.0 - 2024-01-09
 
 ### Added